requirements.txt
editdistance==0.5.2
lmdb==1.0.0
matplotlib==3.2.1
numpy==1.19.5
opencv-python==4.4.0.46
path==15.0.0
tensorflow-gpu==2.4.0